third_party/WebKit/LayoutTests/fast/forms/textarea/textarea-dir-auto-with-placeholder.html

<!DOCTYPE html>
<meta charset="utf-8">
<body>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<textarea dir="auto" placeholder="LTR placeholder string">&#x05E2;</textarea>
<script>
test(() => {
    assert_equals(getComputedStyle(document.querySelector('textarea'), '').direction, 'rtl');
}, 'Placeholder\'s direction shouldn\'t affect TEXTAREA direction.');
</script>
</body>
